refactor(watcher): define a PodWatcher for watcher interface
