+++ pkg/kubernetes/watcher/watcher.go
@@ -11,5 +11,28 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+package watcher
 
-package kubernetes
+import (
+	"context"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/watch"
+	"k8s.io/client-go/kubernetes"
+)
+
+type PodWatcher struct {
+	Core watch.Interface
+}
+
+func NewPodWatcher(c kubernetes.Interface) (*PodWatcher, error) {
+	watcher, err := c.CoreV1().Pods("").Watch(context.TODO(), metav1.ListOptions{})
+	if err != nil {
+		return nil, err
+	}
+
+	return &PodWatcher{
+		Core: watcher,
+	}, nil
+
+}
